@@ -1,5 +1,5 @@
 class CommentsController < ApplicationController
-  before_filter :authenticate_user!, only: [:create, :upvote]
+  before_action :authenticate_user!, only: [:create, :upvote]
 
   def create
     post = Post.find(params[:post_id])
