@@ -1,4 +1,4 @@
 class RecipeListItemsController < ApplicationController
-  before_filter :authenticate_user!, only: [:create, :update]
+  before_action :authenticate_user!, only: [:create, :update]
 
 end
